chore(Utilities): Move REST API v1 into separate folder
