+++ src/Utilities/hooks/api_v1/__fixtures__/supported_ssgs.json
